add AppBarMenu component
