+++ src/connected/AppBar.tsx
@@ -8,24 +8,14 @@ import * as React from "react";
 import { connect } from "react-redux";
 import { SxProps, Theme, styled } from "@mui/material";
 import AppBarComponent from "@mui/material/AppBar";
-import Divider from "@mui/material/Divider";
 import IconButton from "@mui/material/IconButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import Menu from "@mui/material/Menu";
-import MenuItem from "@mui/material/MenuItem";
 import Toolbar from "@mui/material/Toolbar";
 import Tooltip from "@mui/material/Tooltip";
 import Typography from "@mui/material/Typography";
 import CloudDownloadIcon from "@mui/icons-material/CloudDownload";
-import CodeOutlinedIcon from "@mui/icons-material/CodeOutlined";
-import HelpOutlineIcon from "@mui/icons-material/HelpOutline";
-import InfoOutlinedIcon from "@mui/icons-material/InfoOutlined";
 import MoreVertIcon from "@mui/icons-material/MoreVert";
-import SettingsIcon from "@mui/icons-material/Settings";
 import RefreshIcon from "@mui/icons-material/Refresh";
 import ShareIcon from "@mui/icons-material/Share";
-import PolicyIcon from "@mui/icons-material/Policy";
 import { deepOrange } from "@mui/material/colors";
 
 import i18n from "@/i18n";
@@ -35,8 +25,7 @@ import { AppState } from "@/states/appState";
 import { WithLocale } from "@/util/lang";
 import { openDialog } from "@/actions/controlActions";
 import { shareStatePermalink, updateResources } from "@/actions/dataActions";
-import DevRefPage from "@/components/DevRefPage";
-import ImprintPage from "@/components/ImprintPage";
+import AppBarMenu from "@/components/AppBarMenu";
 import UserControl from "./UserControl";
 
 interface AppBarImplProps extends WithLocale {
@@ -128,43 +117,8 @@ const AppBarImpl: React.FC<AppBarImplProps> = ({
   allowSharing,
   shareStatePermalink,
 }: AppBarImplProps) => {
-  const [imprintOpen, setImprintOpen] = React.useState(false);
-  const [devRefOpen, setDevRefOpen] = React.useState(false);
   const [moreMenuOpen, setMoreMenuOpen] = React.useState(false);
   const moreButtonEl = React.useRef<HTMLButtonElement | null>(null);
-
-  const handleSettingsButtonClicked = () => {
-    setMoreMenuOpen(false);
-    openDialog("settings");
-  };
-
-  const handleOpenAbout = () => {
-    setMoreMenuOpen(false);
-    openDialog("about");
-  };
-
-  const handleOpenManual = () => {
-    setMoreMenuOpen(false);
-    window.open("https://xcube-dev.github.io/xcube-viewer/", "Manual");
-  };
-
-  const handleOpenDevRef = () => {
-    setMoreMenuOpen(false);
-    setDevRefOpen(true);
-  };
-
-  const handleCloseDevRef = () => {
-    setDevRefOpen(false);
-  };
-
-  const handleOpenImprint = () => {
-    setMoreMenuOpen(false);
-    setImprintOpen(true);
-  };
-
-  const handleCloseImprint = () => {
-    setImprintOpen(false);
-  };
 
   const handleOpenMoreMenu = () => {
     setMoreMenuOpen(true);
@@ -173,6 +127,7 @@ const AppBarImpl: React.FC<AppBarImplProps> = ({
   const handleCloseMoreMenu = () => {
     setMoreMenuOpen(false);
   };
+
   return (
     <AppBarComponent position="absolute" sx={appBarStyle.appBar} elevation={0}>
       <Toolbar disableGutters sx={styles.toolbar} variant="dense">
@@ -239,50 +194,12 @@ const AppBarImpl: React.FC<AppBarImplProps> = ({
           <MoreVertIcon />
         </IconButton>
       </Toolbar>
-      <ImprintPage open={imprintOpen} onClose={handleCloseImprint} />
-      <DevRefPage open={devRefOpen} onClose={handleCloseDevRef} />
-      <Menu
+      <AppBarMenu
         anchorEl={moreButtonEl.current}
         open={moreMenuOpen}
         onClose={handleCloseMoreMenu}
-      >
-        {Config.instance.branding.allowAboutPage && (
-          <MenuItem onClick={handleOpenAbout}>
-            <ListItemIcon>
-              <InfoOutlinedIcon fontSize="small" />
-            </ListItemIcon>
-            <ListItemText>
-              {i18n.get("About ${appName}", { appName })}
-            </ListItemText>
-          </MenuItem>
-        )}
-        <Divider />
-        <MenuItem onClick={handleOpenManual}>
-          <ListItemIcon>
-            <HelpOutlineIcon fontSize="small" />
-          </ListItemIcon>
-          <ListItemText>{i18n.get("Documentation")}</ListItemText>
-        </MenuItem>
-        <MenuItem onClick={handleOpenDevRef}>
-          <ListItemIcon>
-            <CodeOutlinedIcon fontSize="small" />
-          </ListItemIcon>
-          <ListItemText>{i18n.get("Developer Reference")}</ListItemText>
-        </MenuItem>
-        <MenuItem onClick={handleOpenImprint}>
-          <ListItemIcon>
-            <PolicyIcon fontSize="small" />
-          </ListItemIcon>
-          <ListItemText>{i18n.get("Imprint")}</ListItemText>
-        </MenuItem>
-        <Divider />
-        <MenuItem onClick={handleSettingsButtonClicked}>
-          <ListItemIcon>
-            <SettingsIcon fontSize="small" />
-          </ListItemIcon>
-          <ListItemText>{i18n.get("Settings")}</ListItemText>
-        </MenuItem>
-      </Menu>
+        openDialog={openDialog}
+      />
     </AppBarComponent>
   );
 };
